requirements.txt
beepboop==0.1.1
docutils==0.12
google-api-python-client==1.5.5
lockfile==0.12.2
python-daemon==2.1.1
PyYAML==3.11
requests==2.9.1
six==1.10.0
slackclient==1.0.2
slacker==0.9.9
spacy==1.3.0
stemming=1.0.1
websocket-client==0.35.0
